requirements.txt
pdfplumber
pandas
psycopg2-binary
boto3
python-dotenv